+++ client/src/utils/auth/index.ts
@@ -2,3 +2,4 @@ export { default as addBearer } from '@/utils/auth/addBearer';
 export { default as getAuthorizationHeader } from '@/utils/auth/getAuthorizationHeader';
 export { default as isValidToken } from '@/utils/auth/isValidToken';
 export { default as setAccesstokenToHeader } from '@/utils/auth/setAccesstokenToHeader';
+export { default as loginLocalStorage } from '@/utils/auth/loginLocalStorage';
